Not sure that's really needed, but it makes Stan happy.

+++ app/Console/Commands/VitePublishCommand.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
@@ -34,13 +35,19 @@ class VitePublishCommand extends Command
         $existingFiles = Storage::disk('do')->allFiles('scottzirkel/build');
 
         foreach ($existingFiles as $file) {
+            if (!is_string($file)) {
+                Log::info('File not string? '. json_encode($file));
+                continue;
+            }
             $localFile = str($file)->afterLast('/')->toString();
             $localFile = public_path('/build/assets/'.$localFile);
             if (File::exists($localFile)) {
                 continue;
             }
 
+            $this->info('Deleting '.$file.' from CDN.');
             Storage::disk('do')->delete($file);
+
         }
 
         // do cleanup
